Move tests to extension subfolder
