+++ docker/hipsycl/configure.sh
@@ -8,4 +8,5 @@ cmake . -G Ninja -B build \
     -DCMAKE_BUILD_TYPE=Release \
     -DSYCL_CTS_ENABLE_OPENCL_INTEROP_TESTS=0 \
     -DSYCL_CTS_ENABLE_FULL_CONFORMANCE=0 \
+    -DSYCL_ENABLE_EXT_ONEAPI_PROPERTY_LIST_TESTS=0 \
     $@
