+++ tests/extension/oneapi_property_list/CMakeLists.txt
@@ -1,0 +1,5 @@
+if(SYCL_ENABLE_EXT_ONEAPI_PROPERTY_LIST_TESTS)
+  file(GLOB test_cases_list *.cpp)
+
+  add_cts_test(${test_cases_list})
+endif()
