--- tests/property_list/CMakeLists.txt
@@ -1,3 +1,0 @@
-file(GLOB test_cases_list *.cpp)
-
-add_cts_test(${test_cases_list})
